# backend/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "医院在线预约挂号系统 API",
    "description": "根据当前后端控制器、请求 DTO、测试断言与《项目分工与接口映射.md》整理的可导入 Apifox 的 OpenAPI 3.0 JSON。统一响应格式为 { code, message, data }。\n\n已补充 x-apifox-folder 和 operationId，导入后更容易按模块分组并生成接口结构。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发"
    }
  ],
...
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "注册",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "phone": "13900139000",
                "password": "123456",
                "confirmPassword": "123456",
                "captcha": "8888"
              },
              "examples": {
                "registerSuccess": {
                  "summary": "注册成功",
                  "value": {
                    "phone": "13900139000",
                    "password": "123456",
                    "confirmPassword": "123456",
                    "captcha": "8888"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "integer",
                          "format": "int64",
                          "example": 2
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "registerSuccess": {
                    "summary": "注册成功响应",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "userId": 2
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-apifox-note": "以下 examples 来自现有集成测试，可在 Apifox 中直接作为调试用例。",
        "x-apifox-folder": "认证",
        "operationId": "postApiAuthRegister",
        "x-apifox-summary": "注册"
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "登录",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "phone": "13800138000",
                "password": "123456"
              },
              "examples": {
                "loginSuccess": {
                  "summary": "登录成功",
                  "value": {
                    "phone": "13800138000",
                    "password": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "token-1-123456789"
                        },
                        "userInfo": {
                          "$ref": "#/components/schemas/UserBrief"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "loginSuccess": {
                    "summary": "登录成功响应",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "token": "token-1-123456789",
                        "userInfo": {
                          "id": 1,
                          "name": "张三",
                          "phone": "13800138000",
                          "avatar": "/avatar/user-1.png"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-apifox-folder": "认证",
        "operationId": "postApiAuthLogin",
        "x-apifox-summary": "登录"
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "认证"
        ],
        "summary": "获取当前登录用户",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userInfo": {
                          "$ref": "#/components/schemas/UserBrief"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "meSuccess": {
                    "summary": "获取当前用户成功",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "userInfo": {
                          "id": 1,
                          "name": "张三",
                          "phone": "13800138000",
                          "avatar": "/avatar/user-1.png"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "x-apifox-folder": "认证",
        "operationId": "getApiAuthMe",
        "x-apifox-summary": "获取当前登录用户"
      }
    },
    "/api/auth/change-password": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "修改密码",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              },
              "example": {
                "oldPassword": "123456",
                "newPassword": "abc12345",
                "confirmPassword": "abc12345"
              },
              "examples": {
                "changePasswordSuccess": {
                  "summary": "修改密码成功",
                  "value": {
                    "oldPassword": "123456",
                    "newPassword": "abc12345",
                    "confirmPassword": "abc12345"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "??",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "null",
                      "nullable": true,
                      "example": null
                    }
                  }
                },
                "examples": {
                  "successNull": {
                    "summary": "??",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "x-apifox-folder": "认证",
        "operationId": "postApiAuthChangePassword",
        "x-apifox-summary": "修改密码"
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "退出登录",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "??",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "null",
                      "nullable": true,
                      "example": null
                    }
                  }
                },
                "examples": {
                  "successNull": {
                    "summary": "??",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "x-apifox-folder": "认证",
        "operationId": "postApiAuthLogout",
        "x-apifox-summary": "退出登录"
      }
    },
...
    "/api/consults": {
      "post": {
        "tags": [
          "咨询订单"
        ],
        "summary": "创建咨询订单",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsultRequest"
              },
              "example": {
                "doctorId": 1,
                "familyMemberId": 1,
                "diseaseDesc": "复诊"
              },
              "examples": {
                "createConsult": {
                  "summary": "创建咨询订单",
                  "value": {
                    "doctorId": 1,
                    "familyMemberId": 1,
                    "diseaseDesc": "复诊"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OrderNoResponse"
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "x-apifox-folder": "咨询订单",
        "operationId": "postApiConsults",
        "x-apifox-summary": "创建咨询订单"
      }
    },
    "/api/consults/my": {
      "get": {
        "tags": [
          "咨询订单"
        ],
        "summary": "我的咨询列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageConsultResponse"
                }
              }
            }
          }
        },
        "x-apifox-folder": "咨询订单",
        "operationId": "getApiConsultsMy",
        "x-apifox-summary": "我的咨询列表"
      }
    },
    "/api/consults/{orderNo}": {
      "get": {
        "tags": [
          "咨询订单"
        ],
        "summary": "咨询订单详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orderNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultDetailResponse"
                },
                "examples": {
                  "consultDetail": {
                    "summary": "咨询订单详情",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "orderNo": "CO202607170001",
                        "doctor": {
                          "id": 1,
                          "name": "李主任"
                        },
                        "patientName": "张三",
                        "status": 2,
                        "fee": 39.9,
                        "duration": 15,
                        "appointmentTime": "2026-07-18T10:00:00",
                        "createTime": "2026-07-17T10:00:00"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-apifox-folder": "咨询订单",
        "operationId": "getApiConsultsOrderNo",
        "x-apifox-summary": "咨询订单详情"
      }
    },
    "/api/consults/{orderNo}/pay": {
      "post": {
        "tags": [
          "咨询订单"
        ],
        "summary": "咨询支付",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orderNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayRequest"
              },
              "example": {
                "payType": 1
              },
              "examples": {
                "payConsult": {
                  "summary": "支付咨询订单",
                  "value": {
                    "payType": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "example": true
                        },
                        "payUrl": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "payConsultSuccess": {
                    "summary": "支付成功",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "success": true,
                        "payUrl": null
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-apifox-folder": "咨询订单",
        "operationId": "postApiConsultsOrderNoPay",
        "x-apifox-summary": "咨询支付"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10
        }
      }
    },
    "responses": {
...
      "OrderNoResponse": {
        "description": "成功",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "orderNo": {
                      "type": "string",
                      "example": "AP202607170002"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": [
          "phone",
          "password",
          "confirmPassword"
        ],
        "properties": {
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          },
          "captcha": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "phone",
          "password"
        ],
        "properties": {
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "required": [
          "oldPassword",
          "newPassword",
          "confirmPassword"
        ],
        "properties": {
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        }
      },
...
      "CreateConsultRequest": {
        "type": "object",
        "required": [
          "doctorId",
          "familyMemberId"
        ],
        "properties": {
          "doctorId": {
            "type": "integer",
            "format": "int64"
          },
          "familyMemberId": {
            "type": "integer",
            "format": "int64"
          },
          "diseaseDesc": {
            "type": "string"
          }
        }
      },
      "PayRequest": {
        "type": "object",
        "required": [
          "payType"
        ],
        "properties": {
          "payType": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
      "UserBrief": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "张三"
          },
          "phone": {
            "type": "string",
            "example": "13800138000"
          },
          "avatar": {
            "type": "string",
            "example": "/avatar/user-1.png"
          }
        }
      },
...
      "PageConsultResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "records": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "total": {
                "type": "integer"
              },
              "page": {
                "type": "integer"
              },
              "pageSize": {
                "type": "integer"
              }
            }
          }
        }
      },
      "ConsultDetailResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "orderNo": {
                "type": "string"
              },
              "doctor": {
                "type": "object"
              },
              "patientName": {
                "type": "string"
              },
              "status": {
                "type": "integer"
              },
              "fee": {
                "type": "number"
              },
              "duration": {
                "type": "integer"
              },
              "appointmentTime": {
                "type": "string"
              },
              "createTime": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}